+++ src/main/java/com/curriculo/service/ExperienceService.java
@@ -13,15 +13,23 @@ public class ExperienceService {
     @Autowired
     private ExperienceRepository experienceRepository;
 
-    public List <Experience> getAllExperiences(){
+    // Método para obter todas as experiências
+    public List<Experience> getAllExperiences() {
         return experienceRepository.findAll();
     }
 
-    public Experience saveExperience (Experience experience) {
-        return experienceRepository.save(experience);
+    // Método para salvar uma nova experiência
+    public Experience saveExperience(Experience experience) {
+        return experienceRepository.save(experience); // Salva a experiência e retorna o objeto salvo
     }
 
-    public void deleteExperience(Long id){
-        experienceRepository.deleteById(id);
+    // Método para excluir uma experiência
+    public boolean deleteExperience(Long id) {
+        if (experienceRepository.existsById(id)) { // Verifica se a experiência existe
+            experienceRepository.deleteById(id);
+            return true; // Retorna true se a exclusão for bem-sucedida
+        } else {
+            return false; // Retorna false se não encontrar a experiência
+        }
     }
 }
